# Optimize pooled string array joins