This PR optimizes the [PoolStringArray.join()](https://docs.godotengine.org/en/3.1/classes/class_poolstringarray.html#class-poolstringarray-method-join) function by applying the logic used by the native [StringBuilder](https://github.com/godotengine/godot/blob/master/core/string_builder.cpp#L57) class. **The new implementation is about 9x faster in general and 4x faster for empty delimitiers.**

*Test A*: A single test joins 1000 random strings of 32 characters (GUIDs) with a comma.
*Test B*: A single test joins 1000 random strings of 32 characters (GUIDs) with an empty string.

Each test ran 100 times and these are the average duration of a single run. The time contains the join() call, only.

Test machine: Windows 10 Pro 1903, i7 3.3 GHz, 16 GB, SSD
Custom Build: VC++ 2019, release_debug, x64

| Build | Test A | Test B |
| --- | --- | --- |
|  | Test A | Test B |
| Exported offical 3.1 | 347 ms | 132 ms |
| Editor offical 3.1 | 368 ms | 136 ms |
| Editor w/ old impl | 423 ms | 171 ms |
| **Editor w/ new impl** | 47 ms | 43 ms |
| Editor w/ old StringBuilder | 1013 ms | 563 ms |
| **Editor w/ new StringBuilder** | 174 ms | 107 ms |

All join-results were tested against a static string to be correct.

The test source can be found here: https://gist.github.com/fkollmann/ee72937393fcff74fcd252cdaf2e2ce6