@akien-mga 2019-11-07
Moving to 4.0 milestone as this is not ready to merge yet, and we are now in release freeze for 3.2.

There are still some still issues to address, and commits should be squashed together.

CC @lawnjelly who seems to like testing optimizations :)