@lawnjelly 2019-11-07
I won't be able to profile as I can't work on godot for the next week (and this will probably need profiling from within godot, rather than a testbed, because of the dependencies on string etc), but I've had a quick look through the code, and I can't believe I'm saying this, but I actually agree with most of the author's decisions. :+1: 

The original type of method the author suggested would be the fastest, because although String etc are great for convenience, all the house keeping etc won't beat some optimized code. But I understand the desire to use string builder. The main impetus is to keep the number of allocations down, and I can tell that's the drive behind both designs, with a bulk allocation of chunks, and providing the COW thing works with the strings (I don't know the string implementation well) then that should hopefully stop all the small allocations the previous string builder suffered from.

The previous string builder also suffered from using godot Vector, which does far too many allocations, and hopefully someone will get around to working on the more traditional grow allocation vector alternative that was discussed on IRC a few weeks ago.

I also noted that if you use C style strings with the string builder, you need to keep them around until you get the result as it doesn't copy them (which is sensible, as that would kill performance). That kind of thing that should be written in a comment describing the class. In general personally I find there is far too little comments in godot, but  I don't know the general opinion on that.