@aaronfranke 2020-05-13
@fkollmann Is this still desired? If so, it needs to be rebased on the latest master branch. Since this PR was last updated, PoolArrays have been removed and replaced with PackedArrays. Therefore, this PR may need to be redone from scratch, or may no longer be relevant.

If not, abandoned pull requests will be closed in the future as announced [here](https://godotengine.org/article/headsup-vulkan-merged-master-branch-unstable).