@aaronfranke 2020-09-27
@fkollmann While performance improvements are very welcome, this PR has been inactive for a long time and needs significant work, so I'm closing this. If this is still desired, you are welcome to open another PR.